feat: update

+++ tools-data/src/main/java/com/koala/data/models/netease/detailInfo/SongModel.java
@@ -58,5 +58,5 @@ public class SongModel implements Serializable {
     private String rurl;
     private Integer mst;
     private Integer cp;
-    private Integer publishTime;
+    private Long publishTime;
 }
